Better rounding

+++ api/withdraw.go
@@ -12,7 +12,7 @@ import (
 
 type withdrawJSONRequest struct {
 	Order string  `json:"order"`
-	Sum   float32 `json:"sum"`
+	Sum   float64 `json:"sum"`
 }
 
 func (api *API) HandleWithdraw(w http.ResponseWriter, r *http.Request) {
